Rewrite MergeDrive.FireDrive to support reverse mode
This also updates the supporting code, which needs to be able to use reverse
mode (much like thte same code that's used for the MassDrive). Also
streamlines a few things, like getting rid of extend/retract which are no
longer needed.

The new drive logic has significantly more thrust, since the old logic had a
bug that caused lots of excess downtime.

It appears that this thrusts faster in the reverse direction, which I haven't
figured out yet.