Rewrite InitMergeDrives()
This makes the search for merge drives more general; anything with the pattern
of "merge-rotor-rotor-subgrid" will be recognized as a drive, as long as the
rotors point in the same direction, the merge block points at a right angle to
the rotors, and there's a companion merge block on the main grid to pair with
the subgrid merge block.

It also makes the search faster; it should be O(N) alsmost completely, except
for a minor O(N*M) part checking subgrid merge blocks against main grid merge
blocks. The old code had lots of nested loops and could blow out the time
limit if there were lots of drives.